requirements.txt
numpy
pandas
mysql-connector-python
pymysql
python-dotenv
scikit-learn
seaborn
matplotlib
catboost
xgboost
# -e .